northd: Don't disable console logging for ovn-northd
Many of the OVS utilities (ovs-vsctl and the like) disable logging to
console for good reason, it would confuse and annoy users.  However,
for daemons this makes less sense, as users expect logging to
console if the command line arguments call for it.  This patch brings
ovn-northd into line with the rest of the ovs daemons (ovn-controller,
ovs-vswitchd, ovsdb-server) by removing its manual disabling of
console logging and VLM_reconnect behavior.